refactor(embeddings): purge inline hash-embeddings from published dist (#558)
Completes #558 and unblocks CI on every PR since #545 hardened the
consumer-smoke hash-fallback guard.

Real production call sites replaced with fastembed:
- `hooks-tools.ts` — `generateSimpleEmbedding` for semantic routing and
  the MoE/flash-attention MCP handlers now go through `createEmbeddingService`.
- `benchmarks/pretrain/index.ts` — `benchmarkEmbeddingGeneration` now
  measures the real embedder; the pipeline-sim branch uses random vectors.

Test fixtures relocated under `src/__tests__/` (excluded from publish via
`!**/__tests__/**` in package.json):
- `MockEmbeddingService` → `embeddings/src/__tests__/mocks/mock-embedding-service.ts`.
  `createEmbeddingService` drops the `case 'mock'` branch and throws for
  the string, aligning doc + runtime per #558 AC.
- `MockBatchEmbedder` → `embeddings/src/__tests__/migration/mock-batch-embedder.ts`,
  removed from the migration barrel.

`'mock'` removed from the `EmbeddingProvider` union; `MockEmbeddingConfig`
kept as a standalone interface that the test mock constructs against.

`TestDeterministicEmbedding` (reasoningbank) rewritten as an LCG-seeded
generator so it no longer matches the hash-embedding signature; stays at
its published location since the reasoningbank `useMockEmbeddings` flag
imports it directly.

Consumer-smoke heuristic tightened (`verifyNoInlineHashEmbeddings`):
requires `Float32Array + charCodeAt` AND one of the int→float conversions
— `Math.sin(hash)` within 800 chars, `/ 0xffffffff`, or a direct
`arr[i] = …charCodeAt(…)` assignment. Eliminates 9 false positives
(neural RL algorithms, q-learning router, persistent-cache FNV hashing)
while still catching the minimal-leak fixture from `hash-fallback-guard.test.ts`.

Verified: full suite 7364 pass / 12 skipped, consumer smoke 29/29, lint clean.

Closes #558
